This commit is necessary since c++ code must be object oriented
